fullSchedule: handle innermost zero-dimensional bands
Existing implementation of fullSchedule iteratively takes a flat product
of multi union piecewise affine expressions with partial schedule,
starting from the innermost band.  It also intersects the domain of the
multi union piecewise affine expression with the filter sets.  However,
if the innermost band is zero-dimensional, it is expected to have an
explicit domain, which it does not in this case.  The domain
intersection then fails.  Collect the domain information separately from
taking flat ranges and intersect the domain of the affine expression
only after it was fully constructed.